requirements.txt
Django==3.2.16
psycopg2-binary==2.8.6
python-dotenv==0.20.0
djangorestframework==3.12.4
django-filter==2.4.0
Pillow==9.3.0
drf-base64==2.0